Initial commit logging to database
Adds syncing of tracked programs to database.

+++ backend/src/entities/bookedtime.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, Integer, Float
 from .entity import Entity, Base
 from marshmallow import Schema, fields
 
@@ -7,13 +7,13 @@ class BookedTime(Entity, Base):
     __tablename__ = 'bookedtime'
     task_id = Column(Integer)
     logged_time_id = Column(Integer)
-    booked_time_minutes = Column(Integer)
+    booked_time_minutes = Column(Float)
 
 
 class BookedTimeSchema(Schema):
     id = fields.Integer()
     task_id = fields.Integer()
     logged_time_id = fields.Integer()
-    booked_time_minutes = fields.Integer()
+    booked_time_minutes = fields.Float()
     created_at = fields.DateTime()
     updated_at = fields.DateTime()
